+++ lib/controller/controller.py
@@ -449,12 +449,15 @@ class Controller:
             except KeyboardInterrupt:
                 self.handle_pause()
 
+            time.sleep(0.3)
+
     def add_directory(self, path):
         """Add directory to the recursion queue"""
 
         # Pass if path is in exclusive directories
         if any(
-            "/" + dir in path for dir in options["exclude_subdirs"]
+            path.startswith(dir) or "/" + dir in path
+            for dir in options["exclude_subdirs"]
         ):
             return