+++ lib/parse/cmdline.py
@@ -62,6 +62,13 @@ def parse_arguments():
         help="Load raw HTTP request from file (use '--scheme' flag to set the scheme)",
     )
     mandatory.add_option(
+        "--nmap-report",
+        action="store",
+        dest="nmap_report",
+        metavar="PATH",
+        help="Load targets from nmap report (Ensure the inclusion of the -sV flag during nmap scan for comprehensive results)",
+    )
+    mandatory.add_option(
         "-s", "--session", action="store", dest="session_file", help="Session file"
     )
     mandatory.add_option(
@@ -449,6 +456,8 @@ def parse_arguments():
         help="Number of retries for failed requests",
     )
     connection.add_option("--ip", action="store", dest="ip", help="Server IP address")
+    connection.add_option("--interface", action="store", dest="network_interface", help="Network interface to use")
+
 
     # Advanced Settings
     advanced = OptionGroup(parser, "Advanced Settings")